+++ docs/changelog.rst
@@ -23,11 +23,14 @@ Unreleased
 - :py:mod:`~ws.checkers.WikilinkChecker`: fixed urldecoding of section anchors,
   fixed race condition between the "Archive page" and "Broken section link"
   flags, removed some old and unnecessary workarounds
+- Simplified :py:class:`ws.utils.TLSAdapter` using ``ssl_version`` instead of
+  ``ssl_options``.
 - Many bug fixes for the :py:mod:`ws.db` module and the ``checkdb.py`` script.
 - Migrated to SQLAlchemy 2.0:
   https://docs.sqlalchemy.org/en/20/changelog/migration_20.html
-- Simplified :py:class:`ws.utils.TLSAdapter` using ``ssl_version`` instead of
-  ``ssl_options``.
+- The :py:mod:`ws.db` module now requires SQLAlchemy with ``asyncio`` support
+  and the ``asyncpg`` driver for PostgreSQL. The synchronous interface with the
+  ``psycopg2`` driver is still used as well.
 
 - New scripts: